fix(db): 会话异常/文化合规/连接池预测/错误分析/演练/时区迁移修正（保留字列加引号、AVG(boolean)、示例数据移除、种子按已登记区域过滤）

+++ database/migrations/20260720_210000_create_drill_system.sql
@@ -301,7 +301,7 @@ SELECT
     AVG(rto) AS avg_rto,
     AVG(rpo) AS avg_rpo,
     
-    AVG((results->'sloCompliance'->'availability'->>'passed')::boolean) AS avg_availability_compliance,
+    AVG(((results->'sloCompliance'->'availability'->>'passed')::boolean)::int) AS avg_availability_compliance,
     
     MAX(start_time) AS last_drill_time,
     MIN(start_time) AS first_drill_time
